pyproject.toml
[project]
name = "shiningsoulremake-master"
version = "0.1.0"
description = "Shining Soul remake modernizado com estrutura modular e uv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame>=2.6.1",
    "python-dotenv>=1.2.2",
    "numpy>=1.26",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
